refactor: 전체 경로 수정
