+++ src/components/features/Home/CategorySection/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Link } from 'react-router-dom';
 
-import { useGetCategories } from '@/api/hooks/useGetCategorys';
+import { useGetCategories } from '@/api/hooks/categories/categories.api';
 import { Container } from '@/components/common/layouts/Container';
 import { Grid } from '@/components/common/layouts/Grid';
 import { getDynamicPath } from '@/routes/path';
