+++ src/hooks/useOrderFormContext.ts
@@ -1,5 +1,5 @@
 import { useFormContext } from 'react-hook-form';
 
-import type { OrderFormData } from '@/types';
+import type { OrderFormData } from '@/types/order';
 
 export const useOrderFormContext = useFormContext<OrderFormData>;
